+++ beaker/services/secret.py
@@ -24,7 +24,7 @@ class SecretClient(ServiceClient):
         :raises SecretNotFound: If the secret doesn't exist.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        workspace: Workspace = self._resolve_workspace(workspace)
+        workspace: Workspace = self._resolve_workspace(workspace, read_only_ok=True)
         return Secret.from_json(
             self.request(
                 f"workspaces/{workspace.id}/secrets/{self._url_quote(secret)}",
@@ -49,7 +49,7 @@ class SecretClient(ServiceClient):
         :raises SecretNotFound: If the secret doesn't exist.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
-        workspace: Workspace = self._resolve_workspace(workspace)
+        workspace: Workspace = self._resolve_workspace(workspace, read_only_ok=True)
         name = secret.name if isinstance(secret, Secret) else secret
         return self.request(
             f"workspaces/{workspace.id}/secrets/{self._url_quote(name)}/value",
@@ -70,6 +70,7 @@ class SecretClient(ServiceClient):
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
             :data:`Beaker.config.defeault_workspace <beaker.Config.default_workspace>` are set.
+        :raises WorkspaceWriteError: If the workspace has been archived.
         :raises HTTPError: Any other HTTP exception that can occur.
         """
         workspace: Workspace = self._resolve_workspace(workspace)
@@ -92,6 +93,7 @@ class SecretClient(ServiceClient):
         :raises WorkspaceNotFound: If the workspace doesn't exist.
         :raises WorkspaceNotSet: If neither ``workspace`` nor
             :data:`Beaker.config.defeault_workspace <beaker.Config.default_workspace>` are set.
+        :raises WorkspaceWriteError: If the workspace has been archived.
         :raises SecretNotFound: If the secret doesn't exist.
         :raises HTTPError: Any other HTTP exception that can occur.
         """